feat: Only abort init when dealing with existing files.
